requirements.txt
flask==3.1.3
gunicorn==22.0.0
finnhub-python==2.4.21
yfinance==0.2.51
requests==2.32.4
flask-login>=0.6.3
flask-bcrypt>=1.0.1
flask-limiter>=3.5.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.5
psycopg2-binary>=2.9.9
redis>=5.0.1
prometheus-client>=0.20.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp>=1.24.0
opentelemetry-instrumentation-flask>=0.45b0
opentelemetry-instrumentation-requests>=0.45b0
opentelemetry-instrumentation-sqlalchemy>=0.45b0
python-dotenv==1.0.0
pytest==7.4.4
pytest-cov==4.1.0
wheel>=0.46.2
jaraco.context>=6.1.0
